pyproject.toml
[project]
name = "hospital-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.1",
    "mcp[cli]>=1.16.0",
    "oracledb>=3.4.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
]
